Change up ready generation to be more AXI like
In AXI the 'ready' flags can only be deasserted directly after a successful
transfer. This behavior has been added to the Formal Verification section.

The 'up_rdy' flag is now also reset using the 'rst' signal. Having the control
signals move to a know state during a reset should make debugging easier.

Remove the premature raising of the 'up_rdy' flag. Now 'up_rdy' will be raised
only when the down stream 'dn_rdy' is high or when the up stream has valid data
AND the modules 'down' pipeline has not been stalled. This keeps the behavior
that allowed for data to fall though and be presented on the down stream bus.